requirements.txt
# Core dependencies
akshare>=1.10.0
pymongo>=4.0.0
numpy>=1.20.0
pandas>=1.3.0
ta-lib>=0.4.24

# Backtesting
backtrader>=1.9.76

# Web framework
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
schedule>=1.1.0

# n8n integration
requests>=2.25.0

# Development tools
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0

